% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{BoxPlot}
\alias{BoxPlot}
\title{Boxplot Representation}
\usage{
BoxPlot(x, col = "white", main = "Boxplot representacion", ylab = "",
  legend = TRUE)
}
\arguments{
\item{x}{a numeric vector containing the sample to plot the boxplot representation.}

\item{col}{a single colour to fill the boxplot representation; default to "white".}

\item{main}{a main title for the boxplot; default to "Boxplot representation".}

\item{ylab}{y-axis label; default to empty.}

\item{legend}{logical value; if TRUE (default), details about boxplot representation are given.}
}
\value{
This function is called for the side effect of drawing the plot.
}
\description{
The function \code{BoxPlot} displays a boxplot representation of a given sample.
}
\details{
The quantiles needed to obtain this representation are computed using the function \code{sample.quantile}.
}
\examples{
x=c(5,-5,rnorm(40))
BoxPlot(x,col="pink")
}
