% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesd.R
\name{sample.sd}
\alias{sample.sd}
\title{Sample Standard Deviation}
\usage{
sample.sd(x)
}
\arguments{
\item{x}{a numeric vector containing the sample.}
}
\value{
A single numerical value corresponding with the sample standard deviation.
}
\description{
\code{sample.sd} computes the sample standard deviation of a sample \code{x}.
}
\details{
Given \eqn{\{x_1,\ldots,x_n\}} a sample of a random variable, the sample standard deviation
can be computed as \eqn{S=\sqrt{\frac{1}{n}\sum_{i=1}^n (x_i-\bar{x})^2}}.
}
\examples{
x=rnorm(20)
sample.sd(x)
}
