% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplevar.R
\name{sample.var}
\alias{sample.var}
\title{Sample Variance}
\usage{
sample.var(x)
}
\arguments{
\item{x}{a numeric vector containing the sample.}
}
\value{
A single numerical value corresponding with the sample variance.
}
\description{
\code{sample.var} computes the sample variance of a sample \code{x}.
}
\details{
Given \eqn{\{x_1,\ldots,x_n\}} a sample of a random variable, the sample variance
can be computed as \eqn{S^2=\frac{1}{n}\sum_{i=1}^n (x_i-\bar{x})^2}.
}
\examples{
x=rnorm(20)
sample.var(x)
}
